fix: make location set endpoint system facing (#13194)
* fix: make location set endpoint system facing

* test(events): assert versions array shape on initialisation locations and administrative area seeding
